+++ src/HEPBridgeInsulationTile/HEPBridgeInsulationTile.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{18F0CD86-A2D9-4E41-9509-95A6685F241C}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>HEPWallBridge</RootNamespace>
-    <AssemblyName>HEPWallBridge</AssemblyName>
+    <RootNamespace>HEPBridgeInsulationTile</RootNamespace>
+    <AssemblyName>HEPBridgeInsulationTile</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -16,7 +16,7 @@
   <PropertyGroup>
     <Title>Through-Wall Radbolt Reflector</Title>
     <Description>Adds a new Through-Wall Radbolt Reflector building that receives and redirects Radbolts from Radbolt Generators, through wall and floor tiles without leaking gas or liquid.</Description>
-    <ModVersion>1.1.0</ModVersion>
+    <ModVersion>2.0.0</ModVersion>
     <SupportedContent>EXPANSION1_ID</SupportedContent>
     <UsesPLib>true</UsesPLib>
   </PropertyGroup>
