+++ src/HEPBridgeInsulationTile/HEPWallBridgePatches.cs
@@ -3,15 +3,15 @@ using SanchozzONIMods.Lib;
 using PeterHan.PLib.Core;
 using PeterHan.PLib.PatchManager;
 
-namespace HEPWallBridge
+namespace HEPBridgeInsulationTile
 {
-    internal sealed class HEPWallBridgePatches : KMod.UserMod2
+    internal sealed class HEPBridgeInsulationTilePatches : KMod.UserMod2
     {
         public override void OnLoad(Harmony harmony)
         {
             base.OnLoad(harmony);
             PUtil.InitLibrary();
-            new PPatchManager(harmony).RegisterPatchClass(typeof(HEPWallBridgePatches));
+            new PPatchManager(harmony).RegisterPatchClass(typeof(HEPBridgeInsulationTilePatches));
         }
 
         [PLibMethod(RunAt.BeforeDbInit)]
@@ -23,9 +23,9 @@ namespace HEPWallBridge
         [PLibMethod(RunAt.AfterDbInit)]
         private static void AddBuilding()
         {
-            Utils.AddBuildingToPlanScreen("HEP", HighEnergyParticleWallBridgeRedirectorConfig.ID, HEPBridgeTileConfig.ID);
+            Utils.AddBuildingToPlanScreen("HEP", HEPBridgeInsulationTileConfig.ID, HEPBridgeTileConfig.ID);
             var KleiHEPBridgeTileTech = Db.Get().Techs.TryGetTechForTechItem(HEPBridgeTileConfig.ID)?.Id ?? "NuclearRefinement";
-            Utils.AddBuildingToTechnology(KleiHEPBridgeTileTech, HighEnergyParticleWallBridgeRedirectorConfig.ID);
+            Utils.AddBuildingToTechnology(KleiHEPBridgeTileTech, HEPBridgeInsulationTileConfig.ID);
             PGameUtils.CopySoundsToAnim("wallbridge_orb_transporter_kanim", "orb_transporter_kanim");
         }
     }
