+++ src/HEPBridgeInsulationTile/STRINGS.cs
@@ -1,6 +1,6 @@
 ﻿using STRINGS;
 
-namespace HEPWallBridge
+namespace HEPBridgeInsulationTile
 {
     public class STRINGS
     {
@@ -10,7 +10,7 @@ namespace HEPWallBridge
             {
                 public class HIGHENERGYPARTICLEWALLBRIDGEREDIRECTOR
                 {
-                    public static LocString NAME = UI.FormatAsLink("Through-Wall Radbolt Reflector", HighEnergyParticleWallBridgeRedirectorConfig.ID);
+                    public static LocString NAME = UI.FormatAsLink("Through-Wall Radbolt Reflector", HEPBridgeInsulationTileConfig.ID);
                     public static LocString DESC = "";
                     public static LocString EFFECT = "Receives and redirects Radbolts from " + UI.FormatAsLink("Radbolt Generators", "HIGHENERGYPARTICLESPAWNER") + " through wall and floor tiles without leaking gas or liquid.\n\nFunctions as regular tile.";
                 }
